refactor(impact): centralize all impact logic in one place
