+++ CODEOWNERS
@@ -5,3 +5,4 @@
 /src/app/core/services/stream.service.ts @internxt/cryptography
 /src/app/crypto/services/pgp.service.ts @internxt/cryptography
 /src/upload.worker.ts @internxt/cryptography
+/src/app/analytics/impact.service.ts @internxt/marketing
